Преобразвоание сервиса корзины
